Add a test which breaks spec loading
Currently if you instantiate class which defines hook specs the
introspector (`pluggy.varnames`) breaks on methods and incorrectly
returns the `self` name. I found this by accident. Although it would be
odd it's still conceivable that a user might wish to define hook specs
on an instance. The real problem seems to be that varnames doesn't
actually disregard `self` like the doc string claims.