Deprecate __multicall__ support
Add a new `pluggy.callers._MultiCall` implementation which removes
support the implicit `__multicall__` special argument and drops the
recursion required to handle hook wrappers. Rename the original
implementation `_LegacyMultiCall` and load it only when the
`__multicall__` argument is detected in a hookimpl function
signature. Add a deprecation warning whenever the legacy fallback
occurs.

Resolves #23