Fix #81: Stop using 'inspect.getargspec()'
This function is deprecated and raises the warning.

  DeprecationWarning: inspect.getargspec() is deprecated, use
  inspect.signature() instead

While this suggests we use 'inspect.signature()', there's no reason to
use this instead of the 'inspect.getfullargspec()', which has almost
identical semantics to 'inspect.getargspec()' but handles kwargs-only
functions. For what we want though, this is a drop-in replacement.

Note that because 'inspect.getfullargspec()' isn't provided in Python
2.7, we may see slightly different behavior between Python 2 and 3 (e.g.
for the kwargs-only case above).  There's nothing we can do about this
without vendoring the entire function.

Signed-off-by: Stephen Finucane <stephen@that.guru>