Rework the docs intro and add another example
Rework the beginning of the docs answering three fundamental questions.

Instead of trying to cram everything into a highly compressed role definition, do the following:

Use 4 roles to introduce the basic concept from a bird's eye perspective - 3 of them are programs and one of them is a human (because why not?).

Instead of having only one example have two: the first example becomes
the "toy example" (because that's what it is). Add another example that
is showing two minimal projects (host and plugin), demonstrating the
concepts that would overwhelm the first toy example but are important
to understand conventions and get a complete picture.

Some additional changes in the docs:

* Flesh out the motivation for a system like pluggy a bit more.
* Turn "The pytest plugin system" from a heading into a tag-line.
* fix a pre existing title level problem
* add orphan tag to api reference to be able to build with warnings as errors
* turn doc build warnings into errors
* Made title casing more consistent