Inline _TracedHookExecution
The way _TracedHookExecution is implemented, it takes the PluginManager
instance, which creates a cyclic dependency between the manager.py and
_tracing.py modules. It also mutates internal variables of
PluginManager. This makes the code harder to understand. Just inlining
it makes things mostly straightforward.

This commit also removes an assert which prevented a trace from being
added if there is already an active one. I don't see any reason why it
was done; seems like a legitimate thing to do and should work just fine.