Remove HookSpec - inline into HookCaller
In my opinion, the HookSpec type doesn't pull its weight -- it is
closely tied in a one-to-one relation with HookCaller (possibly set
after initialization), and can be reduced to two attributes (spec_opts
and spec_argnames). So by inlining it into HookCaller, there's one less
type to mentally keep track of.

The type is not public nor mentioned in the API Reference so removing it
is not a breaking change.