hooks: keep hookimpls in a single list
The hookexec receives them in a single list (which is good), so make the
HookCaller keep them in a single array as well, so can avoid the list
concatenation in the hot call path.

This makes adding a hookimpl a little slower, but not by much, and it is
much colder than calling so the tradeoff makes sense. It is possible to
optimize by keeping track of the splitpoint instead of calculating it
every time, but I don't think it's worth it.