result: keep original traceback and reraise with it
The `Result` API allows the same exception to be raised multiple times.
In Python, `Exception.__traceback__` becomes longer everytime an
exception is raised. This means that on every raise, the exception's
traceback gets longer and longer.

To prevent this, save the original traceback and always raise using it.

Regressed in pluggy 1.1.0 (fbc444218c442dd8cbe29bd68cde8fea52b56baf).